Backend/pythonAI/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
python-dotenv>=1.0,<2.0
google-generativeai>=0.7,<1.0
pymongo>=4.6,<5.0
jinja2>=3.1,<4.0
markupsafe>=2.1,<3.1
